Add format arguments checking for log messages
The compiler can check argument formats of printf-like functions.
Note that the compiler may suggest using %z* for size_t arguments but that only applies to macOS 10.13 High Sierra and later so it's best to avoid using %z*. Instead, use %ll* and typecast to (uint64_t) or maybe use a define for the format string and a define for the type cast so it can work with different architectures (for example: PRIKADDR and CASTKADDR)

This change helped find a bug in kern_compression.cpp and also in some other kexts that use Lilu.