Fix early dyld shared cache map load panic
With user patching enabled, Lilu will try to read dyld shared library map. This will panic if it happens too early (before the bsd process for the kernel is set) or fail (if rootvnode hasn't been set by bsd_init).

We need to therefore activate the user patcher later but we still want to activate kernel patcher as early as possible (before certain kexts load).
Therefore, the user patcher and kernel patcher activation needs to be separated.

We can use serial_keyboard_init to know when rootvnode is ready since it happens soon after bsd_init. We can't use mountroot_post_hook because that is only meant to be used by BootCache (it does not chain hooks).
We'll add a function called processUserLoadCallbacks to be called by performCommonInit and by serialKeyboardInit. The first time after both performCommonInit (flagged by userReady) and rootvnode are set is when lilu.processUserLoadCallbacks(userPatcher) will be called.