Add method to set max_protection of vm entries
vm_protect can't add write protection if max_protection of corresponding vm_entry does not include write protection.
vmSetMaxProtection can work around that.
I don't know if this method is useful yet because of other issues blocking certain types of user patches (such as the one that is supposed to be enabled by the -cdfon debug-arg in Big Sur).