Fix UserPatcher::patchBinary
The call to patchSharedCache in patchBinary doesn't do anything if lookupStorage.size is zero. Try fallback to restrict in that case.

Does lookupStorage.size == 0 indicate that restrict will fail? In the case of Big Sur, restrict does fail and the WindowServer process quits. We should not attempt to use restrict if it's going to cause the process to quit. Maybe we only need to check for macOS version ≥ Big Sur to know that restrict will fail? I don't know anything about restrict, but If it means that it will try to load the real library instead of the dyld shared library, then it will definitely fail for Big Sur since there is no real library (at least for the CoreDisplay library that we're wanting to patch in this case).