Call hookMemoryAccess earlier
To catch vm_shared_region_slide which is how we get the slide value (otherwise we might be too late in Big Sur and Monterey to catch the slide value - though I don't know what the slide value is good for currently).

Previously, hookMemoryAccess was called earlier, but then some of the UserPatcher code was moved to after rootvnode was initialized (at serial_keyboard_init time) 76c3477c26d68e5e1149c73b70588ade1762992c .
Maybe that commit should be reverted and somehow we defer file I/O to later when rootvnode is known to be ready (we don't need the info from the files until it's time to make the patches)

But maybe hookMemoryAccess should be here anyway. We can store a flag to know that hookMemoryAccess has succeeded and use it in UserPatcher::registerPatches. The hooked functions can check flags to know when they should do work or do nothing.