End the debug log with some counters
We count the number of times a patch is made and output the info to the debug log before it is written to disk. Useful for debugging to determine how many compares are made (for performance) and how many patches were done (for example, if a WindowServer process CoreDisplay shared library patch happens more than once, then there's probably a problem). This has the added benefit of making it easy to find the end of the debug log as the file is not truncated. Just search for the first occurrence of the counters output.

When performing a patch, include the page offset in the log. This can easily be matched with the page offset in the shared library file to verify that the patch was found in the correct place.