Prefer lookupStorage user patch method
Currently, shared dyld cache patches in BigSur and later don't have lookupStorage info for patches (e.g. CoreDisplay framework for WindowServer process), but other user patches do have lookupStorage info (e.g. displaypolicyd process). The difference is because Lilu doesn't have code to load the binary for shared dyld cache frameworks. We would want the code to load only the part of the dyld shared cache containing the framework to be patched but that would require code to parse the dyld shared cache file.

Anyway, the non-lookupStorage patch method should be attempted only after the lookupStorage patch method, because the lookupStorage patch method requires that a page is identical to the contents of the binary file before doing a patch - it will not perform a patch if the page was modified by a different patcher.

If we're able to do a lookupStorage patch, then don't try a non-lookupStorage patch.