Wrap some functions for debugging
For now we'll just count how many times the functions are called.
Code sign functions in the kernel are used for doing user patches when the patch is not for a framework in the dyld shared cache or the framework in the cache does not exist outside the cache, such as in Big Sur and later.
Specifically, cs_validate_page is used, except in certain macOS versions where cs_validate_page doesn't exist and cs_validate_range is used instead.
Counters are used as a sanity check - to see how often each function is called which may indicate an issue if the counts are not as expected.

vm_page_validate_cs_mapped calls vm_page_validate_cs_mapped_slow which calls cs_validate_page. There are other callers of vm_page_validate_cs_mapped_slow that we don't count. If count of vm_page_validate_cs_mapped is zero, then it means those other callers are being used and that's ok/expected.