Fix cs_validate_page wrapper for BigSur
It is a void function, not a boolean function.
Also, improve the names of the counters and add a couple more.
Everything should be fine if cs_validate_page is called ≈100000 times in the first 4 minutes after boot, and counterPagePatchWithoutLookupStorageBigSur is called almost as many times.
counterPagePatchReplaceSuccess counts the replacements that use the lookupStorage method (e.g. displaypolicyd).
counterPagePatchWithoutLookupStorageSearchReplaceSuccess counts replacements that don't use lookupStorage info (e.g. CoreDisplay framework used by WindowServer in Big Sur or. later).